Domain model supports user connections, encpoints are secured

+++ our-schedule/rest-schedule/src/main/java/com/arianit/rest/schedule/service/AppointmentService.java
@@ -1,9 +1,11 @@
 package com.arianit.rest.schedule.service;
 
 import com.arianit.rest.schedule.entity.Appointment;
+import com.arianit.rest.schedule.entity.User;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
+import jakarta.ws.rs.BadRequestException;
 
 import java.util.List;
 
@@ -12,22 +14,28 @@ import java.util.List;
 public class AppointmentService {
 
     @Transactional(Transactional.TxType.SUPPORTS)
-    public List<Appointment> findAllAppointments() {
-        return Appointment.listAll();
+    public List<Appointment> findAllAppointments(Long userId) {
+        return Appointment.findAppointments(userId);
     }
 
     @Transactional(Transactional.TxType.SUPPORTS)
-    public Appointment findAppointmentById(Long id) {
-        return Appointment.findById(id);
+    public Appointment findAppointmentById(Long appointmentId, Long userId) {
+        return Appointment.findAppointment(appointmentId, userId);
     }
 
-    public Appointment persistAppointment(@Valid Appointment appointment) {
+    public Appointment persistAppointment(User user, @Valid Appointment appointment) {
+        if (!user.equals(appointment.getUser())) {
+            throw new BadRequestException("User ID in the appointment does not match the authenticated user.");
+        }
         appointment.persist();
         return appointment;
     }
 
-    public Appointment updateAppointment(@Valid Appointment appointment) {
+    public Appointment updateAppointment(User user, @Valid Appointment appointment) {
         Appointment entity = Appointment.findById(appointment.id);
+        if (!user.equals(entity.getUser())) {
+            throw new BadRequestException("User ID in the appointment does not match the authenticated user.");
+        }
         entity.title = appointment.title;
         entity.description = appointment.description;
         entity.date = appointment.date;
@@ -36,8 +44,11 @@ public class AppointmentService {
         return entity;
     }
 
-    public void deleteAppointment(Long id) {
+    public void deleteAppointment(User user, Long id) {
         Appointment appointment = Appointment.findById(id);
+        if (!user.equals(appointment.getUser())) {
+            throw new BadRequestException("User ID in the appointment does not match the authenticated user.");
+        }
         appointment.delete();
     }
 }
